requirements.txt
typer[all]==0.9.0
pydantic==2.6.0
PyYAML==6.0.1
rich==13.7.0
